requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv

# djongo
# pymongo

# mongoengine 
# django-mongoengine

supabase
dj-database-url